+++ src/backend/efiling-graphql/src/main/java/ca/bc/gov/open/jag/submission/ReviewResource.java
@@ -1,8 +1,7 @@
 package ca.bc.gov.open.jag.submission;
 
-import ca.bc.gov.ag.csows.filing.status.File;
-import ca.bc.gov.open.jag.efilingcommons.model.FilePackage;
-import ca.bc.gov.open.jag.efilingcommons.service.EfilingStatusService;
+import ca.bc.gov.open.jag.efilingcommons.submission.EfilingStatusService;
+import ca.bc.gov.open.jag.efilingcommons.submission.models.FilingPackage;
 import ca.bc.gov.open.jag.submission.model.FilePackageOutput;
 import org.eclipse.microprofile.graphql.Description;
 import org.eclipse.microprofile.graphql.GraphQLApi;
@@ -32,15 +31,13 @@ public class ReviewResource {
     public FilePackageOutput getSubmission(@Name("clientId") BigDecimal clientId, @Name("packageNo") BigDecimal packageNo) {
         logger.info("GraphQl Request received");
 
-        Optional<FilePackage> filePackage = efilingStatusService.findStatusByPackage(clientId, packageNo);
+        Optional<FilingPackage> filePackage = efilingStatusService.findStatusByPackage(clientId, packageNo);
 
         FilePackageOutput filePackageOutput = new FilePackageOutput();
 
         if (filePackage.isPresent()) {
-            filePackageOutput.setCourtClass(filePackage.get().getCourtClassCd());
-            filePackageOutput.setFileNumber(filePackage.get().getClientFileNo());
+            filePackageOutput.setCourtClass(filePackage.get().getApplicationCode());
         }
-
 
         return filePackageOutput;
     }
