+++ src/backend/libs/efiling-commons/src/main/java/ca/bc/gov/open/jag/efilingcommons/service/EfilingSubmissionService.java
@@ -1,7 +1,7 @@
 package ca.bc.gov.open.jag.efilingcommons.service;
 
 import ca.bc.gov.open.jag.efilingcommons.model.AccountDetails;
-import ca.bc.gov.open.jag.efilingcommons.model.FilingPackage;
+import ca.bc.gov.open.jag.efilingcommons.submission.models.FilingPackage;
 import ca.bc.gov.open.jag.efilingcommons.model.SubmitPackageResponse;
 
 public interface EfilingSubmissionService {
